Fix BOOTIMG_FIT for new AOSP restrictions
Make the BOOTIMG_FIT target work correctly with current AOSP. The
current AOSP build system blocks the use of mkimage because it calls out
to 'dtc', which may not have been build as a host tool, and uses the
host system's copy, which is no longer allowed.

This change makes sure the dependencies are property specified for 'dtc'
and ensures PATH is set correctly for the invocation by mkimage.

Test: local build for beagle_x15 lunch target
Change-Id: Ie1e7df43b96054f736cd0ce194220078dda74ef0
Signed-off-by: Alistair Strachan <astrachan@google.com>