Fix a race in BOOTIMG_FIT Makefile
The rules for BOOTIMG_FIT depended on INSTALLED_KERNEL_TARGET and
INSTALLED_RAMDISK_TARGET, but these variables may not have been defined
at the point the Makefile was parsed and simply-expanded, so the rules
to build BOOTIMG_FIT could race with the creation of the installed kernel
and ramdisk targets.

Fix this problem by locally defining BOOTIMG_FIT_INSTALLED_KERNEL_TARGET
and BOOTIMG_FIT_INSTALLED_RAMDISK_TARGET to sensible defaults and copy
them to PRIVATE_ variants in the recipe.

Bug: 110964307
Test: local build with m -j128
Change-Id: I49a107c08a3e010035260c21258a7c7dade6c7e7
Signed-off-by: Alistair Strachan <astrachan@google.com>